add util file
